Apply next.js link component to bottom navigation bar
Former-commit-id: 7cc3495be222e4f44d061d5628bedbcbf84ac7da
