+++ nugurang-front/package.json
@@ -14,7 +14,7 @@
     "@material-ui/system": "^4.9.14",
     "apollo-client": "^2.6.10",
     "clsx": "^1.1.1",
-    "next": "^9.5.3",
+    "next": "^9.5.4",
     "next-redux-wrapper": "^6.0.2",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
