+++ nugurang-front/pages/team/index.js
@@ -1,0 +1,7 @@
+import TeamInfo from './team_info'
+
+export default function Index() {
+  return (
+      <TeamInfo />
+  );
+}
